requirements.txt
langchain_chroma>=0.1.4
langchain_community>=0.3.10
langchain_core>=0.3.22
langchain_openai>=0.2.11
langchain_text_splitters>=0.3.2
openai>=1.55.3
pydub>=0.25.1
PyPDF>=3.1.0
python-dotenv==1.0.1
gtts==2.5.4
PyPDF2==3.0.1
fastapi==0.115.7
edge-tts==7.0.0
python-multipart==0.0.20
google-generativeai==0.8.4
langchain-google-genai==2.0.9